% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{oaxaca_blinder_svy}
\alias{oaxaca_blinder_svy}
\title{Oaxaca-Blinder Decomposition with Bootstrap Confidence Intervals using svyglm}
\usage{
oaxaca_blinder_svy(formula, data, group, weights, R = 1000, conf.level = 0.95)
}
\arguments{
\item{formula}{An object of class formula: the model formula.}

\item{data}{A data frame: containing the variables in the formula, group, and weights.}

\item{group}{A character string: the name of the binary group variable in the data. 1 for the first group, and 0 for the second group.}

\item{weights}{A character string: the name of the weights variable in the data.}

\item{R}{An integer: the number of bootstrap replicates. Default is 1000.}

\item{conf.level}{A float less than 1: Confidence level for estimated CI.}
}
\value{
A list containing the mean and confidence intervals for endowments, coefficients, and interaction components.
}
\description{
Oaxaca-Blinder Decomposition with Bootstrap Confidence Intervals using svyglm
}
\examples{
\dontrun{
# Assuming a dataset 'data' with response 'y', predictors 'x1' and 'x2', a binary group 'group', and weights 'w'
result <- oaxaca_blinder_svy(y ~ x1 + x2, data = data, group = "group", weights = "w", R = 1000)
}
}
